016. Authorization with Gates.mp4

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -8,12 +8,19 @@ use App\Http\Traits\CanLoadRelationships;
 use App\Models\Attendee;
 use App\Models\Event;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class AttendeeController extends Controller
 {
     use CanLoadRelationships;
 
     private array $relations = ['user'];
+
+    public function __construct()
+    {
+        //this protect every methods with middleware and connect to routes
+        $this->middleware('auth:sanctum')->except(['index', 'show', 'update']);
+    }
 
     /**
      * Display a listing of the resource.
@@ -59,8 +66,10 @@ class AttendeeController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $event, Attendee $attendee)
+    public function destroy(Event $event, Attendee $attendee)
     {
+        Gate::authorize('delete-attendee', [$event, $attendee]);
+
         $attendee->delete();
 
         return response(status: 204);
